fastapi-taskman/requirements.txt
aiosqlite==0.21.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astroid==3.3.9
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.2
dill==0.4.0
Django==5.2
dnspython==2.6.1
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==33.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
packaging==25.0
passlib==1.7.4
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
pylint==3.3.7
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.1
rsa==4.9.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sqlparse==0.5.3
starlette==0.38.5
tomli==2.2.1
tomlkit==0.13.2
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
